+++ CHANGELOG.md
@@ -6,6 +6,18 @@ rename `Unreleased` topic with the new version tag. Finally, create a new `Unrel
 
 ## Unreleased
 
+## v9.4.0
+
+### DevOps: new test suite
+- Initiate a new Node.js-based test suite.
+
+### Bugfixes/Improvements
+- Clean project files if the `neu create` command wan't successful.
+- Display an error message if the `neu build --copy-storage` fails due to missing `.storage` directory.
+- Support Darwin `arm64` binaries natively without binary translation (`x64` -> `arm64`).
+- Make the client library downloading process optional and support importing the client library via `@neutralinojs/lib`.
+- Download the module version of the client library based on the app configuration.
+
 ## v9.3.1
 
 ### Core: Plugins